requirements.txt
selenium==4.18.1
pytest==8.3.5
allure-pytest==2.13.5
allure-python-commons==2.13.5
webdriver-manager==4.0.1
Faker==37.1.0
